fix: history's handling of backspaces & arrows inside words & macros
It used to be that macro_in_paren would corrupt the history, because the
key_left would immediately apply to the previous word, before the new entry
containing the parens was pushed onto the history stack. Now we push the entry
as soon as it needs to be incremented for the first time, so arrows and
backspaces apply to it.

You can still get kinda weird history behavior if you try hard enough. Try
putting enough arrows or backspaces to move outside of the current entry, and it
will start directly editing other entries without splitting them first.