feat: draft new mode-switching commands (breaking change)
Include the mode_enum in the SWITCH_TO command's sequence, as an argument.

This changes the command list in the settings file, and the names of the
switching commands in the kmap files!

Messy! Should get cleaned up soon as part of the refactoring for
internationalization, which will force us to handle c enums in KeyPresses
properly.